+++ stroom-ui/src/components/AppSearchBar/redux.js
@@ -1,17 +1,13 @@
-import { createActions, combineActions, handleActions } from 'redux-actions';
+import { createActions, handleActions } from 'redux-actions';
 
 import { updateIdSubstate } from 'lib/reduxFormUtils';
 
 const actionCreators = createActions({
-  OPEN_DROPDOWN: pickerId => ({ pickerId, isOpen: true }),
-  CLOSE_DROPDOWN: pickerId => ({ pickerId, isOpen: false }),
   SWITCH_MODE: (pickerId, mode) => ({ pickerId, mode }),
   NAVIGATE_TO_FOLDER: (pickerId, navFolder) => ({ pickerId, navFolder }),
   SEARCH_TERM_UPDATED: (pickerId, searchTerm) => ({ pickerId, searchTerm }),
   SEARCH_RESULTS_RETURNED: (pickerId, searchResults) => ({ pickerId, searchResults }),
 });
-
-const { openDropdown, closeDropdown } = actionCreators;
 
 const SEARCH_MODE = {
   GLOBAL_SEARCH: 0,
@@ -23,7 +19,6 @@ const defaultPickerState = {
   searchTerm: '',
   searchResults: [],
   navFolder: undefined,
-  isOpen: false,
   searchMode: SEARCH_MODE.NAVIGATION,
 };
 
@@ -31,10 +26,6 @@ const defaultState = {};
 
 const reducer = handleActions(
   {
-    [combineActions(openDropdown, closeDropdown)]: (state, { payload: { pickerId, isOpen } }) =>
-      updateIdSubstate(state, pickerId, defaultPickerState, {
-        isOpen,
-      }),
     SWITCH_MODE: (state, action) =>
       updateIdSubstate(state, action.payload.pickerId, defaultPickerState, {
         searchMode: action.payload.mode,