+++ stroom-ui/src/components/DocRefInfoModal/DocRefInfoModal.js
@@ -16,10 +16,11 @@
 import React from 'react';
 
 import { connect } from 'react-redux';
-import { compose, branch, renderNothing, renderComponent } from 'recompose';
+import { compose, branch, renderNothing, renderComponent, withProps } from 'recompose';
 
-import { Modal, Button, Form, Loader } from 'semantic-ui-react';
+import { Header, Button, Form, Loader } from 'semantic-ui-react';
 
+import ThemedModal from 'components/ThemedModal';
 import { actionCreators } from './redux';
 
 const { docRefInfoClosed } = actionCreators;
@@ -37,13 +38,24 @@ const enhance = compose(
     ({ docRefInfo }) => !docRefInfo,
     renderComponent(() => <Loader active>Awaiting DocRef Info </Loader>),
   ),
+  withProps(({ docRefInfo: { createTime, updateTime } }) => ({
+    formattedCreateTime: new Date(createTime).toLocaleString('en-GB', { timeZone: 'UTC' }),
+    formattedUpdateTime: new Date(updateTime).toLocaleString('en-GB', { timeZone: 'UTC' }),
+  })),
 );
 
-const DocRefInfoModal = ({ isOpen, docRefInfo, docRefInfoClosed }) => (
-  <Modal open={isOpen} onClose={docRefInfoClosed} size="small" dimmer="inverted">
-    {' '}
-    <Modal.Header>Doc Ref Info</Modal.Header>
-    <Modal.Content scrolling>
+const DocRefInfoModal = ({
+  isOpen,
+  docRefInfo,
+  docRefInfoClosed,
+  formattedCreateTime,
+  formattedUpdateTime,
+}) => (
+  <ThemedModal
+    open={isOpen}
+    onClose={docRefInfoClosed}
+    header={<Header className="header" icon="info" content="Document Information" />}
+    content={
       <Form>
         <Form.Group widths="equal">
           <Form.Input label="Type" type="text" value={docRefInfo.docRef.type} />
@@ -52,29 +64,21 @@ const DocRefInfoModal = ({ isOpen, docRefInfo, docRefInfoClosed }) => (
         </Form.Group>
         <Form.Group widths="equal">
           <Form.Input label="Created by" type="text" value={docRefInfo.createUser} />
-          <Form.Input
-            label="at"
-            type="text"
-            value={new Date(docRefInfo.createTime).toLocaleString('en-GB', { timeZone: 'UTC' })}
-          />
+          <Form.Input label="at" type="text" value={formattedCreateTime} />
         </Form.Group>
         <Form.Group widths="equal">
           <Form.Input label="Updated by" type="text" value={docRefInfo.updateUser} />
-          <Form.Input
-            label="at"
-            type="text"
-            value={new Date(docRefInfo.updateTime).toLocaleString('en-GB', { timeZone: 'UTC' })}
-          />
+          <Form.Input label="at" type="text" value={formattedUpdateTime} />
         </Form.Group>
         <Form.Input label="Other Info" type="text" value={docRefInfo.otherInfo} />
       </Form>
-    </Modal.Content>
-    <Modal.Actions>
+    }
+    actions={
       <Button negative onClick={docRefInfoClosed}>
         Close
       </Button>
-    </Modal.Actions>
-  </Modal>
+    }
+  />
 );
 
 export default enhance(DocRefInfoModal);