+++ stroom-ui/src/lib/withSelectableItemListing/withSelectableItemListing.js
@@ -76,28 +76,28 @@ const withSelectableItemListing = propsFunc =>
         openItem,
         enterItem,
         goBack,
-        selectableItemListing: { selectedItems, focussedItem, selectionBehaviour },
+        selectableItemListing,
         keyIsDown,
       }) => (e) => {
-        if (e.key === 'ArrowUp' || (e.ctrlKey && e.key === 'k')) {
+        if (e.key === 'ArrowUp' || e.key === 'k') {
           focusUp(listingId);
           e.preventDefault();
-        } else if (e.key === 'ArrowDown' || (e.ctrlKey && e.key === 'j')) {
+        } else if (e.key === 'ArrowDown' || e.key === 'j') {
           focusDown(listingId);
           e.preventDefault();
         } else if (e.key === 'Enter') {
-          if (focussedItem) {
-            openItem(focussedItem);
+          if (selectableItemListing.focussedItem) {
+            openItem(selectableItemListing.focussedItem);
           }
           e.preventDefault();
-        } else if (e.ctrlKey && (e.key === 'ArrowRight' || e.key === 'l')) {
-          if (focussedItem) {
-            enterItem(focussedItem);
+        } else if (e.key === 'ArrowRight' || e.key === 'l') {
+          if (selectableItemListing.focussedItem) {
+            enterItem(selectableItemListing.focussedItem);
           }
-        } else if (e.ctrlKey && (e.key === 'ArrowLeft' || e.key === 'h')) {
-          goBack();
+        } else if (e.key === 'ArrowLeft' || e.key === 'h') {
+          goBack(selectableItemListing.focussedItem);
         } else if (e.key === ' ') {
-          if (selectionBehaviour !== SELECTION_BEHAVIOUR.NONE) {
+          if (selectableItemListing.selectionBehaviour !== SELECTION_BEHAVIOUR.NONE) {
             selectFocussed(listingId, keyIsDown);
             e.preventDefault();
           }