+++ stroom-ui/src/sections/AppChrome/appChromeRoutes.js
@@ -11,7 +11,6 @@ import DataViewer from 'sections/DataViewer';
 import UserSettings from 'sections/UserSettings';
 import PathNotFound from 'components/PathNotFound';
 import IFrame from 'components/IFrame';
-import AppSearch from 'sections/AppSearch';
 
 const renderWelcome = props => <AppChrome activeMenuItem="Welcome" content={<Welcome />} />;
 
@@ -82,11 +81,6 @@ export default [
     exact: true,
     path: '/s/apikeys',
     render: props => <AppChrome activeMenuItem="API Keys" content={<ApiTokensIFrame />} />,
-  },
-  {
-    exact: true,
-    path: '/s/search',
-    render: props => <AppChrome activeMenuItem="Search" content={<AppSearch />} />,
   },
   {
     exact: true,